+++ source/renderer/app/containers/wallet/dialogs/WalletImportDialogContainer.js
@@ -13,6 +13,10 @@ type Props = InjectedProps;
 export default class WalletImportDialogContainer extends Component<Props> {
   static defaultProps = { actions: null, stores: null };
 
+  onOpen = () => {
+    this.props.actions.walletMigration.resetMigration.trigger();
+  };
+
   onConfirm = () => {
     this.props.actions.walletMigration.nextStep.trigger();
   };
@@ -54,6 +58,7 @@ export default class WalletImportDialogContainer extends Component<Props> {
             exportSourcePath={exportSourcePath}
             exportErrors={exportErrors}
             pendingImportWalletsCount={pendingImportWalletsCount}
+            onOpen={this.onOpen}
             onConfirm={this.onConfirm}
             onClose={this.onCancel}
             onOpenExternalLink={openExternalLink}